+++ mapchete/io/_path.py
@@ -1,0 +1,114 @@
+import os
+from urllib.request import urlopen
+from urllib.error import HTTPError
+
+from mapchete.io._misc import get_boto3_bucket
+
+
+def path_is_remote(path, s3=True):
+    """
+    Determine whether file path is remote or local.
+
+    Parameters
+    ----------
+    path : path to file
+
+    Returns
+    -------
+    is_remote : bool
+    """
+    prefixes = ("http://", "https://", "/vsicurl/")
+    if s3:
+        prefixes += ("s3://", "/vsis3/")
+    return path.startswith(prefixes)
+
+
+def path_exists(path):
+    """
+    Check if file exists either remote or local.
+
+    Parameters
+    ----------
+    path : path to file
+
+    Returns
+    -------
+    exists : bool
+    """
+    if path.startswith(("http://", "https://")):
+        try:
+            urlopen(path).info()
+            return True
+        except HTTPError as e:
+            if e.code == 404:
+                return False
+            else:
+                raise
+    elif path.startswith("s3://"):
+        bucket = get_boto3_bucket(path.split("/")[2])
+        key = "/".join(path.split("/")[3:])
+        for obj in bucket.objects.filter(Prefix=key):
+            if obj.key == key:
+                return True
+        else:
+            return False
+    else:
+        return os.path.exists(path)
+
+
+def absolute_path(path=None, base_dir=None):
+    """
+    Return absolute path if path is local.
+
+    Parameters
+    ----------
+    path : path to file
+    base_dir : base directory used for absolute path
+
+    Returns
+    -------
+    absolute path
+    """
+    if path_is_remote(path):
+        return path
+    else:
+        if os.path.isabs(path):
+            return path
+        else:
+            if base_dir is None or not os.path.isabs(base_dir):
+                raise TypeError("base_dir must be an absolute path.")
+            return os.path.abspath(os.path.join(base_dir, path))
+
+
+def relative_path(path=None, base_dir=None):
+    """
+    Return relative path if path is local.
+
+    Parameters
+    ----------
+    path : path to file
+    base_dir : directory where path sould be relative to
+
+    Returns
+    -------
+    relative path
+    """
+    if path_is_remote(path) or not os.path.isabs(path):
+        return path
+    else:
+        return os.path.relpath(path, base_dir)
+
+
+def makedirs(path):
+    """
+    Silently create all subdirectories of path if path is local.
+
+    Parameters
+    ----------
+    path : path
+    """
+    if not path_is_remote(path):
+        try:
+            os.makedirs(path)
+        except OSError:
+            pass
